🚚 api: move certificationReport model to src directory
