+++ api/tests/unit/infrastructure/serializers/jsonapi/certification-serializer_test.js
@@ -1,7 +1,7 @@
 import { expect, EMPTY_BLANK_AND_NULL, domainBuilder } from '../../../../test-helper.js';
 import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/certification-serializer.js';
 import { WrongDateFormatError } from '../../../../../lib/domain/errors.js';
-import { NO_EXAMINER_COMMENT } from '../../../../../lib/domain/models/CertificationReport.js';
+import { NO_EXAMINER_COMMENT } from '../../../../../src/certification/shared/domain/models/CertificationReport.js';
 import { CertificationCourse } from '../../../../../lib/domain/models/CertificationCourse.js';
 
 describe('Unit | Serializer | JSONAPI | certification-serializer', function () {
